+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-09T18:01:04.261Z",
+    "2026-06-10T05:17:44.336Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {